+++ ZwiftActivityMonitorV2/nsis-setup/ZwiftActivityMonitorV2-Full.nsi
@@ -2,7 +2,7 @@
 Unicode True
 
 !define PRODUCT_NAME "Zwift Activity Monitor Pro"
-!define PRODUCT_VERSION "2.0.5"
+!define PRODUCT_VERSION "2.1.0"
 !define PRODUCT_PUBLISHER "Kevin Ruff p/b EnJoy Fitness"
 !define PRODUCT_WEB_SITE "https://github.com/ruffk/ZwiftActivityMonitor"
 !define PRODUCT_DIR_REGKEY "Software\Microsoft\Windows\CurrentVersion\App Paths\ZwiftActivityMonitorV2.exe"
