+++ ZwiftActivityMonitorV2/src/extensions/EncryptingJsonConverter.cs
