+++ ZwiftActivityMonitorV2/src/extensions/GroupBoxEx.cs
