+++ ZwiftActivityMonitorV2/src/extensions/MSoffice2010ColorManager.cs
