+++ ZwiftActivityMonitorV2/src/extensions/NotifyPropertyChangedBase.cs
