+++ ZwiftActivityMonitorV2/src/extensions/PropertyInfoHelper.cs
