+++ ZwiftActivityMonitorV2/src/extensions/SyncBindingSource.cs
